# Hash VecDeque in its slice parts