@bluss 2016-01-26
Benchmark code: https://gist.github.com/bluss/a021d1a71b2709dbf724