@Gankro 2016-01-26
Can you add a test that verifies that different splits yield the same hash? Something like pushpushpush == pushpushpoppushpush