@Gankro 2016-01-26
Also: dang! Should we be doing this for eq/ord, etc?