@bluss 2016-01-26
I'm working on this for eq. We should absolutely use slices everywhere we can (llvm likes them, see #30805).

A test to verify same hash, same eq for different splits sounds good. Will do.