@bluss 2016-01-26
I added a test!