@bluss 2016-01-26
Thanks for the feedback Gankro! Together we will hash these bytes, faster than ever.